+++ model/database/transactions/delt_transaction.py
@@ -7,7 +7,7 @@ class DeltTransaction:
     def __init__(self,
                  delt_id: int = 0,
                  user_id: int = 0,
-                 tickets: int = 0,
+                 tickets: float = 0,
                  time: str = None,
                  description: str = None,
                  type_: TransactionType = TransactionType.unknown):
