+++ model/database/transactions/tpay_transaction.py
@@ -6,8 +6,8 @@ class TpayTransaction:
                  tpay_id: int = 0,
                  sender_id: int = 0,
                  receiver_id: int = 0,
-                 transfer: int = 0,
-                 fee: int = 0,
+                 transfer: float = 0,
+                 fee: float = 0,
                  time: str = None,
                  description: str = None):
         self.tpay_id = tpay_id
